added support for document multi-attachments

+++ modules/documentdel.php
@@ -3,7 +3,7 @@
 /*
  * LMS version 1.11-git
  *
- *  (C) Copyright 2001-2013 LMS Developers
+ *  (C) Copyright 2001-2016 LMS Developers
  *
  *  Please, see the doc/AUTHORS for more information about authors!
  *
@@ -24,65 +24,58 @@
  *  $Id$
  */
 
-if($_GET['is_sure']=='1')
-{
-        if(isset($_POST['marks']))
-	{
-	        foreach($_POST['marks'] as $id => $mark)
-		{
-			$md5sum = $DB->GetOne('SELECT c.md5sum FROM documentcontents c
-				JOIN documents d ON (d.id = c.docid)
+if ($_GET['is_sure'] == '1') {
+	if (isset($_POST['marks'])) {
+		foreach($_POST['marks'] as $id => $mark) {
+			$md5sums = $DB->GetCol('SELECT a.md5sum FROM documentattachments a
+				JOIN documents d ON (d.id = a.docid)
 				JOIN docrights r ON (r.doctype = d.type)
-				WHERE c.docid = ? AND r.userid = ? AND (r.rights & 16) = 16',
+				WHERE a.docid = ? AND r.userid = ? AND (r.rights & 16) = 16',
 				array($id, $AUTH->id));
 
-			if (!$md5sum)
+			if (!$md5sums)
 				continue;
 
-			if($DB->GetOne('SELECT COUNT(*) FROM documentcontents WHERE md5sum = ?',array((string)$md5sum))==1)
-			{
-				$filename_pdf = DOC_DIR . DIRECTORY_SEPARATOR . substr($md5sum,0,2) . DIRECTORY_SEPARATOR . $md5sum.'.pdf';
-				if(file_exists($filename_pdf))
-					@unlink($filename_pdf);
-				@unlink(DOC_DIR . DIRECTORY_SEPARATOR . substr($md5sum,0,2) . DIRECTORY_SEPARATOR . $md5sum);
-			}
-	
+			foreach ($md5sums as $md5sum)
+				if ($DB->GetOne('SELECT COUNT(*) FROM documentattachments WHERE md5sum = ?', array((string)$md5sum)) == 1) {
+					$filename_pdf = DOC_DIR . DIRECTORY_SEPARATOR . substr($md5sum,0,2) . DIRECTORY_SEPARATOR . $md5sum.'.pdf';
+					if(file_exists($filename_pdf))
+						@unlink($filename_pdf);
+					@unlink(DOC_DIR . DIRECTORY_SEPARATOR . substr($md5sum,0,2) . DIRECTORY_SEPARATOR . $md5sum);
+				}
+
 			$DB->BeginTrans();
-			
+
 			$DB->Execute('DELETE FROM documentcontents WHERE docid = ?',array($id));
 			$DB->Execute('DELETE FROM documents WHERE id = ?',array($id));
-	
+
 			$DB->CommitTrans();
 		}
-	}
-	elseif(isset($_GET['id']))
-	{			
-		$md5sum = $DB->GetOne('SELECT c.md5sum FROM documentcontents c
-			JOIN documents d ON (d.id = c.docid)
+	} elseif(isset($_GET['id'])) {
+		$md5sums = $DB->GetCol('SELECT a.md5sum FROM documentattachments a
+			JOIN documents d ON (d.id = a.docid)
 			JOIN docrights r ON (r.doctype = d.type)
-			WHERE c.docid = ? AND r.userid = ? AND (r.rights & 16) = 16',
+			WHERE a.docid = ? AND r.userid = ? AND (r.rights & 16) = 16',
 			array($_GET['id'], $AUTH->id));
 
-		if (!$md5sum)
-		{
+		if (!$md5sums) {
 			$SMARTY->display('noaccess.html');
-		        die;
+			die;
 		}
-		
-		if($DB->GetOne('SELECT COUNT(*) FROM documentcontents WHERE md5sum = ?',array((string)$md5sum))==1)
-		{
-			@unlink(DOC_DIR.'/'.substr($md5sum,0,2).'/'.$md5sum);
-		}
-	
+
+		foreach ($md5sums as $md5sum)
+			if ($DB->GetOne('SELECT COUNT(*) FROM documentattachments WHERE md5sum = ?', array((string)$md5sum)) == 1)
+				@unlink(DOC_DIR.'/'.substr($md5sum,0,2).'/'.$md5sum);
+
 		$DB->BeginTrans();
-		
+
 		$DB->Execute('DELETE FROM documentcontents WHERE docid = ?',array($_GET['id']));
 		$DB->Execute('DELETE FROM documents WHERE id = ?',array($_GET['id']));
-	
+
 		$DB->CommitTrans();
 	}
 }
 
 $SESSION->redirect('?'.$SESSION->get('backto'));
-	
+
 ?>
